requirements.txt
gliner>=0.2.0
#torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
tabulate>=0.9.0
accelerate>=0.20.0
datasets